Use sonOfHoward; move to local computer to use IDE
The code now uses sonOfHoward, though I'm not entirely sure where
the code exists on it's development. It seems like it is almost
finished with path_1_stat.

It will now be moved over to the local machine to use pycharm.
It's current state will be assessed and hopefully path_1_stat
can be completed. Documentation will also be updated to reflect
the progress of the code.