Changed plotting program and fixed compatability
Previously the code used pyqtgraph to make plots of peaks, since
it was thought that this module would be quicker than matplotlib.
Unfortunately this module locks up the computer when run, since
it tries to open a plotting window for each plot before deleting
each plotting window. To remedy this, the module PyGnuplot
was implemented instead.

There was also an issue with using the code on hugoniot. This was
due to a difference in numpy versions on hugoniot versus the
laptop that was used for development. The laptop was updated to
the current numpy version to try and fix this. The cause of the
issue was the way the function numpy.isclose works; previously
this would work fine in an "if, else" statement, but now something
has gone wrong. As a result numpy.close has been dropped and now
a different control statement is used (the length of a vector is
now checked to be less than 0.0001, rather than checking if it is
close to 0.0).