Updated documentation, fixed "run_soh" bug
The documentation for the code has been updated to reflect the
current progress of the code.

There was a bug which meant that if the code was ran with "run_soh = False"
the code would fall over. This was because the code would delete
all of the data in the beginning to give it a fresh start (intended to
prevent data from previous runs piling up), which is incompatible
with the "run_soh = False" option. The code will now only delete
the data if "run_soh = True".