Changed module for reading in soh outputs
It was found that numpy.loadtxt was falling over when peakfinder
was being used on other machines. To negate this, pandas is now
being used. Pandas is also quicker. Another change that was made
was the addition of some status messages so that temperature is
now output to the screen when it is calculated.

The kx ky kz debye temperature stuff still needs cleaning up.