Fix reading .ft files; Fix Deby-Waller in kx,ky,kz
The library used to read data from .ft files was previously numpy,
but this was removed due to compatability issues between hugoniot
and my laptop. Numpy was replaced with pandas, but this has some
unexpected consequences which broke the code. This issue has now
been resolved.

There was also an issue with determining the Debye-Waller effect
in certain directions. This was due to the wrong pos list being
used as input to the calc_debye_waller.py function.