Fix compression ratio in Debye-Waller calc
It turns out that the compression ratios were hard coded to 1, 1, 1
which resulted in there being no difference in temperature estimates.
It also resulted in a severe underestimation of the temperature based
on the slope of ln(I) vs G^2. This has now been fixed so that it
uses the compression ratios calculated earlier in the code.

There is a big mismatch between the estimated temperatures from
each model, but one of them generates an accurate temperature.