Add ability to change soh command; format changes
The ability to change the soh run command has been added so that
the code can be run on a supercomputer, and the command to run soh
is changed via the input script. This will speed up the analysis.
Some formatting of the velocity distribution  graph has been
changed also.

The velocity distribution graph should contain a Maxwell-Boltzmann
curve to illustrate how well equilibrated the atoms are.